requirements.txt
pandas
pyarrow
azure-identity    
azure-storage-blob
duckdb
requests